+++ db/migrate/20180328195929_add_language_code_to_sheet_transactions.rb
@@ -1,0 +1,5 @@
+class AddLanguageCodeToSheetTransactions < ActiveRecord::Migration[5.2]
+  def change
+    add_column :sheet_transactions, :language_code, :string
+  end
+end